Add support for user avatars downloaded from avatar providers.
Attempt to download user avatar (from RoboHash for automatically generated
avatars, or from Gravatar via redirection if the user has created an avatar
there) when the user visits a page. The MD5 hash of the user's email address
is used as the ID for looking up avatars.

Cache avatar in local storage for fast access to the image when moving between
pages. Update the version of the image in local storage whenever a more
up-to-date version is downloaded from the avatar provider (which is checked
when first loading a page or when switching between pages). Avatars may also be
cached automatically by the browser based on the header properties of the
response from the avatar provider.

Add crypto library and React sync hooks library to support the managing of
avatars.