[#148] Add support for end-to-end testing with Playwright.
Move local development Dockerfile from using a node base image to using a
Playwright base image (Ubuntu-based rather than Debian).

Add security profile to the docker-compose flow (to support testing with
Chromium).

Add npm packages to support running Playwright with Jest and TypeScript.

Add end-to-end test directory `e2e` with Jest and TypeScript configs and a
basic end-to-end test (loads the front page and accepts cookies). Test
generates a video of the operations peformed and a screenshot at the end of the
test.

Add configurations for testing on Chromium, Firefox and Webkit on a mix of
approximated mobile, laptop and desktop devices.

Add `test:e2e` target to the `yarn` and `run_docker.sh` commands.

Rename `test` target to `test:unit` for the `yarn` and `run_docker.sh` commands.

Update READMEs based on the above changes.

Add cleanup code to the local development `dev.sh` script to ensure that
processes are cleanly stopped when the container is stopped.