Add Material UI support.
Initial implementation that updates parts of the front page, including the map
collection.

Make `Navigation` component a separate "permanent" route that isn't recreated
when changing page. Pages can set the title in the navbar via a prop that holds
the currently displayed title, and a prop that alters the title.